+++ Formula/sqitch.rb
@@ -1,58 +1,8 @@
 require 'formula'
-require 'pp'
+require_relative '../requirements/perl510_req'
+require_relative '../requirements/snowflake_req'
 
 class Sqitch < Formula
-  class Perl510Req < Requirement
-    fatal true
-
-    satisfy(build_env: false) { `perl -E 'print $]'`.to_f >= 5.01000 }
-
-    def message
-      "Sqitch requires Perl 5.10.0 or greater."
-    end
-  end
-
-  class SnowflakeReq < Requirement
-    @@dylib = "/opt/snowflake/snowflakeodbc/lib/universal/libSnowflake.dylib"
-    @@binary = "/Applications/SnowSQL.app/Contents/MacOS/snowsql"
-    @snowsql = false
-    @driver = false
-    fatal true
-    download "https://docs.snowflake.net/manuals/user-guide/odbc-mac.html"
-
-    def initialize(tags = [])
-      super
-      @name = "Snowflake"
-      @driver = File.exist?(@@dylib)
-      @snowsql = File.executable?(@@binary)
-    end
-
-    satisfy(build_env: false) do
-      @driver # Require the driver to build.
-    end
-
-    def message
-      <<~EOS
-        Sqitch Snowflake support requires the Snowflake ODBC driver.
-        Installation: #{ download }
-      EOS
-    end
-
-    def notes
-      msg = "Snowflake support requires the Snowflake ODBC driver and SnowSQL\n\n" \
-            "- Found ODBC driver #{ @@dylib }\n"
-      if @snowsql
-        msg += "- Found SnowSQL binary: #{ @@binary }\n" \
-               "  Make sure it's in your \$PATH or tell Sqitch where to find it by running:\n\n" \
-               "      sqitch config --user engine.snowflake.client #{ @@binary }\n"
-      else
-        msg += "- SnowSQL not found; installation instructions:\n" \
-               "  https://docs.snowflake.net/manuals/user-guide/snowsql-install-config.html\n"
-      end
-      return msg
-    end
-  end
-
   homepage   'https://sqitch.org/'
   version    '0.9998'
   url        "http://cpan.cpantesters.org/authors/id/D/DW/DWHEELER/App-Sqitch-#{stable.version}.tar.gz"
@@ -157,6 +107,7 @@ class Sqitch < Formula
   end
 
   def post_install
+    # Show notes from requirements.
     requirements.each do |req|
       if req.class.method_defined? :notes
         ohai "#{ req.name } Support Notes", req.notes, "\n"
